refactor: 여행지 form부분 리팩토링
