+++ src/pages/Admin/Trip/components/AddDestModal.jsx
@@ -1,6 +1,6 @@
 import styles from "./AddDestModal.module.css";
-import FormAddCity from "./FormAddCity";
-import FormAddCountry from "./FormAddCountry";
+import FormCity from "./FormCity";
+import FormCountry from "./FormCountry";
 
 const AddDestModal = ({content}) => {
 
@@ -9,9 +9,9 @@ const AddDestModal = ({content}) => {
             <p className={styles.title}>{content} 추가하기</p>
             <div className={styles.content_container}>
                 {content==="나라" ? (
-                <FormAddCountry /> 
+                <FormCountry /> 
                 ) : ( 
-                <FormAddCity /> 
+                <FormCity /> 
                 )}
             </div>
         </div>
